% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qicpack-package.R
\docType{data}
\name{dietox}
\alias{dietox}
\title{Dietox data set}
\format{A data frame with 861 observations on 7 variables.
\describe{
\item{Pig}{Identifier}
\item{Weight}{Weight}
\item{Feed}{the cumulated feed intake}
\item{Time}{Time (in weeks) in the experiment}
\item{Evit}{3 different levels of vitamin E supplement}
\item{Cu}{3 different levles of copper supplement}
\item{Litter}{Identifier of litter of each pig} 
}}
\source{
Lauridsen, C., Højsgaard, S., Sørensen, M.T.C. (1999) Influence of Dietary Rapeseed Oli, Vitamin E, and Copper on Performance and Antioxidant and Oxidative Status of Pigs. J. Anim. Sci.77:906-916
}
\usage{
data(dietox)
}
\description{
This data set gives weight of pigs measured weekly for 12 weeks in a 3x3
factorial experiment. Data also contains the starting weight (i.e. the weight at week
1). The treatments are 3 different levels of Evit = vitamin E (dose: 0, 100, 200 mg 
dl-alpha-tocopheryl acetat /kg feed) in combination with 3 different levels of Cu
=copper (dose: 0, 35, 175 mg/kg feed) in the feed. The cumulated feed intake is
 also recorded. The pigs are littermates.
}
\details{
The dietox data frame has 861 observations on 7 variables. This dataset is included 
in multiple packages, including \code{doBy} and \code{geepack}.

#'
}
\examples{
## For examples see example(QIC)
}
\keyword{datasets}
