% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qic.R
\name{QIC}
\alias{QIC}
\alias{get_QIC}
\title{Calculates QIC (Pan, 2001) for model generated with geeglm in the geepack package}
\usage{
QIC(object, ...)

QIC(object, ...)

get_QIC(object)
}
\arguments{
\item{object}{an object class "geeglm", obtained from a call to \code{geeglm} of the \code{geepack} package.}

\item{...}{optionally more fitted model objects.}
}
\value{
If just one object is provided, a numeric value with the corresponding QIC. 

If multiple objects are provided, a data.frame with rows corresponding to the
 objects and columns representing the QIC, log of quasi-likelihood 
 (\code{log.QLik}), trace or the penalty term (\code{Trace}) as well as the number
  of covariates (\code{px}) in the model. 

\code{get_QIC} is the function that does the actual calculation.
}
\description{
This function calculates the quasi-likelihood under the independence model citerion
(also commonly known as quasi-likelihood information criterion) (QIC; Pan, 2001) for
model generated using geeglm in geepack.
}
\details{
The QIC can be used as the AIC for generalized estimating equations. The smaller the
QIC, the better the fit. 

According to Cui and Qian (2007), QICu is obtinaed by assuming the asymptotic
 equivalence of the working correlation and the independence correlation structures.
 It follows that QICu is not appropriate for being used to select an optimal 
 correlation structure and therefore QICu is not included here. 

Based on the full model including all explanatory variables and their interactions, 
the best correlation structure is usually selected first based on the model with the
smallest QIC value. Based on the best correlation structure, we can then further
select the best subset of covariates. The model with the smallest QIC value at this 
stage would be consider as the most parsimonious model with the best correlation
structure.
}
\examples{
data(dietox)
dietox$Cu <- as.factor(dietox$Cu)
gee01 <- geeglm(Weight ~ Time + Cu + Cu * Time, id =Pig, data =
 dietox, family=gaussian,corstr="ex")
gee02 <- geeglm(Weight ~ Time + Cu + Cu * Time, id =Pig, data =
 dietox, family=gaussian,corstr="unstructured")
QIC(gee01)
QIC(gee01, gee02)

mf3 = formula(Weight ~ Cu + Time + I(Time^2))
gee3.ar = geeglm(mf3, data = dietox, id = Pig, family = Gamma, corstr = "ar1")
gee3.i = update(gee3.ar, corstr = "independence")
gee3.ex = update(gee3.ar, corstr = "exchangeable")
gee3.un = update(gee3.ar, corstr = "unstructured")
QIC(gee3.ar, gee3.i, gee3.ex, gee3.un)
}
\references{
Cui, J, and Qian, G. (2007). Selection of Working Correlation Structure and Best
Model in GEE Analyses of Longitudinal Data. \emph{Communication in Statistics
--Simulation and Computation} \bold{36}, 987--996.

Pan, W. (2001). Akaike's Information Criterion in Generalized Estimating Equations.
\emph{Biometrics} \bold{57}, 120--125.
}
\author{
Originally by Daniel J. Hocking, more recent revisions by Thomas Fung.
}
