\name{QICpack-package}
\alias{QICpack-package}
\alias{QICpack}
\docType{package}
\title{
What the package does (short line)
Package for calculating QIC (Pan 2001)
}
\description{
More about what it does (maybe more than one line)
Package contains functions to calculate QIC and compare generalized estimating equation models based on QIC.
}
\details{
\tabular{ll}{
Package: \tab QICpack\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2012-12-10\cr
License: \tab GPL (>= 3)\cr
}
qic(model.R) calculates QIC for a single model fit with geeglm in geepack

qictab(cand.set, modnames) calculates QIC values for a list of models and compares the models based on QIC.
}
\author{
Daniel J. Hocking

Maintainer: Daniel J. Hocking <dhocking@unh.edu>
}
\references{
Burnham, K. P. and D. R. Anderson. 2002. Model selection and multimodel inference: a practical information-theoretic approach. Second edition. Springer Science and Business Media, Inc., New York.
Pan, W. 2001. Akaike's information criterion in generalized estimating equations. Biometrics 57:120-125.

}

\keyword{ package }
\seealso{

}
\examples{

}
