\name{QICpack-package}
\alias{QICpack-package}
\alias{QICpack}
\docType{package}
\title{
Model selection for generalized estimating equations using QIC
}
\description{
This package contains the function 'qic', which calculates Pan's (2001) quasi likelihood information criteria (QIC) for models generated from the gee pack R package. The package also contains a function 'qictab' that compares a list of geeglm models from geepack and calculates delta QIC, QIC model weights, and cumulative model weights.
}
\details{
\tabular{ll}{
Package: \tab QICpack\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-12-10\cr
License: \tab MIT License\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Daniel J. Hocking <dhocking@unh.edu>

Maintainer: Daniel J. Hocking <dhocking@unh.edu>
}
\references{
Burnham, K. P. and D. R. Anderson. 2002. Model selection and multimodel inference: a practical information-theoretic approach. Second edition. Springer Science and Business Media, Inc., New York.

Hardin, J. W. and J. M. Hilbe. 2003. Generalized estimating equations. Chapman and Hall, New York.

Hojsgaard, S., U. Halekoh, and J. Yan. 2006. The R package geepack for generalized estimating equations. Journal of Statistical Software 15:1-11.

Pan, W. 2001. Akaike's information criterion in generalized estimating equations. Biometrics 57:120-125.
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
