% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qicpack-package.R
\docType{package}
\name{qicpack-package}
\alias{qicpack-package}
\alias{qicpack}
\title{qicpack: QIC Package for GEE}
\description{
qicpack: QIC Package for GEE
}
\references{
Fung, T. (2019). \emph{qicpack}: QIC Package for GEE. R package version 0.0.1.
Hocking, D.J. (2016). \emph{qicpack}: qicpack R package. \url{https://github.com/djhocking/qicpack}.
}
\keyword{package}
